Move inner records out of TransportVersionUtils (#132872)
The utilities class has basically become a package itself. This commit
moves inner records from TransportversionUtils and some related
functions out to their own files.
